move intervals into common
